+++ tinted-builder-rust/tests/operation_build.rs
@@ -3,7 +3,7 @@ mod test_utils;
 use anyhow::{Context, Result};
 use std::fs;
 use std::path::PathBuf;
-use test_utils::{run_command, write_to_file};
+use test_utils::{copy_dir_all, run_command, write_to_file};
 
 fn setup(system: &str, scheme_name: &str) -> Result<(String, String, String, String)> {
     let config_file_path: PathBuf =
@@ -275,9 +275,6 @@ fn test_operation_build_base24() -> Result<()> {
         base24_template_rendered_content_fixture,
     ) = setup(system, scheme_name)?;
 
-    if themes_path.is_dir() {
-        fs::remove_dir_all(&themes_path)?;
-    }
     if template_theme_path.is_dir() {
         fs::remove_dir_all(&template_theme_path)?;
     }
@@ -355,9 +352,6 @@ fn test_operation_build_mixed() -> Result<()> {
         base24_template_rendered_content_fixture,
     ) = setup("base24", base24_scheme_name)?;
 
-    if themes_path.is_dir() {
-        fs::remove_dir_all(&themes_path)?;
-    }
     if template_theme_path.is_dir() {
         fs::remove_dir_all(&template_theme_path)?;
     }
@@ -415,6 +409,74 @@ fn test_operation_build_mixed() -> Result<()> {
         ),
         "stdout does not contain the exptected output"
     );
+
+    Ok(())
+}
+
+#[test]
+fn test_operation_build_listbase16() -> Result<()> {
+    // -------
+    // Arrange
+    // -------
+    let name = "operation_build_list";
+    let template_theme_path = PathBuf::from(format!("./template-{}", name));
+    let template_templates_path = template_theme_path.join("templates");
+    let schemes_path = template_theme_path.join("schemes");
+    let rendered_list_theme_path = template_theme_path.join("list-list.md");
+    let rendered_listbase16_theme_path = template_theme_path.join("listbase16-list.md");
+    let rendered_listbase24_theme_path = template_theme_path.join("listbase24-list.md");
+
+    if template_theme_path.is_dir() {
+        fs::remove_dir_all(&template_theme_path)?;
+    }
+    fs::create_dir_all(&template_templates_path)?;
+    fs::copy(
+        "./tests/fixtures/templates/list-config.yaml",
+        template_theme_path.join("templates/config.yaml"),
+    )?;
+    fs::copy(
+        "./tests/fixtures/templates/list-template.mustache",
+        template_theme_path.join("templates/list.mustache"),
+    )?;
+    copy_dir_all("./tests/fixtures/schemes", &schemes_path)?;
+
+    // ---
+    // Act
+    // ---
+    let (stdout, stderr) = run_command(vec![
+        "build".to_string(),
+        template_theme_path.display().to_string(),
+        format!("--schemes-dir={}", schemes_path.display()),
+    ])
+    .unwrap();
+    let rendered_list_content = fs::read_to_string(rendered_list_theme_path)?;
+    let rendered_listbase16_content = fs::read_to_string(rendered_listbase16_theme_path)?;
+    let rendered_listbase24_content = fs::read_to_string(rendered_listbase24_theme_path)?;
+    let expected_list_content = fs::read_to_string("./tests/fixtures/rendered/list.md")?;
+    let expected_listbase16_content =
+        fs::read_to_string("./tests/fixtures/rendered/listbase16.md")?;
+    let expected_listbase24_content =
+        fs::read_to_string("./tests/fixtures/rendered/listbase24.md")?;
+
+    // ------
+    // Assert
+    // ------
+    assert_eq!(rendered_list_content, expected_list_content);
+    assert_eq!(rendered_listbase16_content, expected_listbase16_content);
+    assert_eq!(rendered_listbase24_content, expected_listbase24_content);
+    assert!(
+        stderr.is_empty(),
+        "stderr does not contain the expected output"
+    );
+    let expected_output = format!(
+        r#"Successfully generated "list" list with filename "{0}/{{{{ scheme-system }}}}-list.md"
+Successfully generated "listbase16" list with filename "{0}/{{{{ scheme-system }}}}-list.md"
+Successfully generated "listbase24" list with filename "{0}/{{{{ scheme-system }}}}-list.md"
+"#,
+        template_theme_path.display(),
+    );
+
+    assert_eq!(stdout, expected_output);
 
     Ok(())
 }